use of coverage unified
git-svn-id: svn+ssh://svn.msr-inria.inria.fr/var/lib/svn/repository/tla/trunk/tla2-inria@13684 76a6fc44-f60b-0410-a9a8-e67b0e8fc65c

+++ tlatools/src/tlc2/TLC.java
@@ -260,7 +260,7 @@ public class TLC
                 {
                     try
                     {
-                        TLCGlobals.coverageInterval = (int) (Float.parseFloat(args[index]) * 60 * 1000);
+                        TLCGlobals.coverageInterval = Integer.parseInt(args[index]) * 60 * 1000;
                         if (TLCGlobals.coverageInterval < 0)
                         {
                             printErrorMsg("Error: expect a nonnegative integer for -coverage option.");
